Rewrite sqf to be separate module
