+++ server/package.json
@@ -17,7 +17,7 @@
     "vscode-languageserver-textdocument": "^1.0.2"
   },
   "devDependencies": {
-    "@types/node": "^14.17.32",
+    "@types/node": "^20.11.17",
     "@typescript-eslint/eslint-plugin": "^6.21.0",
     "@typescript-eslint/parser": "^6.21.0",
     "eslint": "^8.56.0",
